# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Todo:
    properties:
      details:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gerenciar tarefas (todos) com Go e SQLite.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo API
  version: "1.0"
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todas as tarefas cadastradas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Lista todas as tarefas
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Adiciona uma nova tarefa ao sistema
      parameters:
      - description: Nova Tarefa
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Cria uma nova tarefa
      tags:
      - todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Remove uma tarefa pelo ID fornecido
      parameters:
      - description: ID da Tarefa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta uma tarefa
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Busca uma tarefa pelo ID
      parameters:
      - description: ID da Tarefa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retorna uma tarefa específica
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Atualiza os detalhes de uma tarefa pelo ID
      parameters:
      - description: ID da Tarefa
        in: path
        name: id
        required: true
        type: integer
      - description: Atualização da Tarefa
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza uma tarefa existente
      tags:
      - todos
swagger: "2.0"